Assignment05/pyproject.toml
[project]
name = "lab06"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.4",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "wandb>=0.25.1",
    "datasets>=2.18.0",
    "pyyaml>=6.0",
]
